% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reports.R
\name{add_go_enrichment_to_report}
\alias{add_go_enrichment_to_report}
\title{Add GO Enrichment Summary to Dynamic Report}
\usage{
add_go_enrichment_to_report(con, project_id, go_content_file, verbose = TRUE)
}
\arguments{
\item{con}{Database connection}

\item{project_id}{Project ID}

\item{go_content_file}{Path to GO report content RDS file}

\item{verbose}{Logical. Print progress information}
}
\description{
Add GO Enrichment Summary to Dynamic Report
}
