% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{check_uniprot_connection}
\alias{check_uniprot_connection}
\title{Updated connection check for annotate_blast_results function}
\usage{
check_uniprot_connection(verbose = TRUE)
}
\description{
Place this code within your annotate_blast_results function
to replace the existing connection check
}
