% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{close_funseq_db}
\alias{close_funseq_db}
\title{Close a database connection}
\usage{
close_funseq_db(con, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
Logical. TRUE if the connection was successfully closed, FALSE otherwise.
}
\description{
Close a database connection
}
\keyword{internal}
