% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect_funseq_db}
\alias{connect_funseq_db}
\title{Connect to an existing funseqR database}
\usage{
connect_funseq_db(db_path, verbose = TRUE)
}
\arguments{
\item{db_path}{Character string specifying the path to the database file.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A connection object to the database.
}
\description{
This function connects to an existing funseqR database.
}
