% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_schema.R
\name{create_funseq_schema}
\alias{create_funseq_schema}
\title{Database schema definition for funseqR}
\usage{
create_funseq_schema(con, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
Invisible NULL.
}
\description{
This file contains functions to create and manage the schema of the funseqR database.
}
\details{
Create the funseqR database schema

This function creates the necessary tables and indexes for the funseqR database.
}
