% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_visualization.R
\name{create_go_comparison_plot}
\alias{create_go_comparison_plot}
\title{Create multi-ontology comparison plot}
\usage{
create_go_comparison_plot(
  bp_results = NULL,
  mf_results = NULL,
  cc_results = NULL,
  top_n = 8
)
}
\arguments{
\item{bp_results}{Data frame. Biological Process enrichment results}

\item{mf_results}{Data frame. Molecular Function enrichment results}

\item{cc_results}{Data frame. Cellular Component enrichment results}

\item{top_n}{Integer. Number of top terms per category. Default is 8}
}
\value{
ggplot object showing comparison across ontologies
}
\description{
Create multi-ontology comparison plot
}
