% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_visualization.R
\name{create_interactive_go_plot}
\alias{create_interactive_go_plot}
\title{Create interactive GO enrichment plot using plotly}
\usage{
create_interactive_go_plot(
  enrichment_results,
  max_terms = 25,
  min_fold_enrichment = 1.5
)
}
\arguments{
\item{enrichment_results}{Data frame. Output from perform_go_enrichment()}

\item{max_terms}{Integer. Maximum number of terms to show. Default is 25}

\item{min_fold_enrichment}{Numeric. Minimum fold enrichment to display. Default is 1.5}
}
\value{
plotly object (interactive plot)
}
\description{
Create interactive GO enrichment plot using plotly
}
