% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{create_project}
\alias{create_project}
\title{Project management functions for funseqR}
\usage{
create_project(
  con,
  project_name,
  description = NULL,
  create_report = TRUE,
  report_format = "Rmd",
  report_template = "standard",
  verbose = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{project_name}{Character string specifying the name of the project.}

\item{description}{Character string describing the project. Default is NULL.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
The ID of the newly created project.
}
\description{
These functions manage projects within the funseqR database.
}
\details{
Create a new project in the database

This function creates a new project in the funseqR database.
}
