% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project from the database}
\usage{
delete_project(con, project_id, confirm = TRUE, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{project_id}{The ID of the project to delete.}

\item{confirm}{Logical. If TRUE, user confirmation will be required. Default is TRUE.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
Logical. TRUE if the deletion was successful.
}
\description{
This function deletes a project and all associated data from the database.
}
