% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_workflow.R
\name{.add_kegg_annotations_to_stored_locus_table}
\alias{.add_kegg_annotations_to_stored_locus_table}
\title{Add KEGG annotations to stored locus table}
\usage{
.add_kegg_annotations_to_stored_locus_table(con, locus_data, verbose)
}
\description{
Add KEGG annotations to stored locus table
}
\keyword{internal}
