% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.add_kegg_brite_classifications}
\alias{.add_kegg_brite_classifications}
\title{Add KEGG BRITE classifications using KEGGREST}
\usage{
.add_kegg_brite_classifications(kegg_data, verbose)
}
\description{
Add KEGG BRITE classifications using KEGGREST
}
\keyword{internal}
