% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.add_kegg_functional_modules}
\alias{.add_kegg_functional_modules}
\title{Add functional module assignments}
\usage{
.add_kegg_functional_modules(kegg_data, verbose)
}
\description{
Add functional module assignments
}
\keyword{internal}
