% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_visualization.R
\name{.apply_y_transform}
\alias{.apply_y_transform}
\title{Apply y-axis transformation}
\usage{
.apply_y_transform(values, transform = "-log10")
}
\arguments{
\item{values}{Numeric vector of values to transform}

\item{transform}{Character. Transformation type}
}
\value{
Transformed numeric vector
}
\description{
Apply y-axis transformation
}
\keyword{internal}
