% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.batch_insert_annotations}
\alias{.batch_insert_annotations}
\title{Batch insert annotations}
\usage{
.batch_insert_annotations(con, annotations)
}
\description{
Batch insert annotations
}
\keyword{internal}
