% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.batch_insert_blast_results}
\alias{.batch_insert_blast_results}
\title{Batch insert BLAST results}
\usage{
.batch_insert_blast_results(con, blast_results)
}
\description{
Batch insert BLAST results
}
\keyword{internal}
