% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.batch_insert_candidate_loci}
\alias{.batch_insert_candidate_loci}
\title{Batch insert candidate loci}
\usage{
.batch_insert_candidate_loci(con, candidates)
}
\description{
Batch insert candidate loci
}
\keyword{internal}
