% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.batch_insert_flanking_sequences}
\alias{.batch_insert_flanking_sequences}
\title{Batch insert flanking sequences}
\usage{
.batch_insert_flanking_sequences(con, flanking_data)
}
\description{
Batch insert flanking sequences
}
\keyword{internal}
