% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.batch_insert_functional_annotations}
\alias{.batch_insert_functional_annotations}
\title{Batch insert functional annotations}
\usage{
.batch_insert_functional_annotations(con, table_name, func_data)
}
\description{
Batch insert functional annotations
}
\keyword{internal}
