% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.batch_insert_locus_statistics}
\alias{.batch_insert_locus_statistics}
\title{Batch insert locus statistics}
\usage{
.batch_insert_locus_statistics(con, locus_stats)
}
\description{
Batch insert locus statistics
}
\keyword{internal}
