% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.batch_insert_reference_sequences}
\alias{.batch_insert_reference_sequences}
\title{Batch insert reference sequences}
\usage{
.batch_insert_reference_sequences(con, ref_sequences)
}
\description{
Batch insert reference sequences
}
\keyword{internal}
