% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.build_eggnog_landscape_query}
\alias{.build_eggnog_landscape_query}
\title{Build eggNOG landscape query}
\usage{
.build_eggnog_landscape_query(blast_param_id)
}
\description{
Build eggNOG landscape query
}
\keyword{internal}
