% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.build_interpro_landscape_query}
\alias{.build_interpro_landscape_query}
\title{Build InterPro landscape query}
\usage{
.build_interpro_landscape_query(blast_param_id)
}
\description{
Build InterPro landscape query
}
\keyword{internal}
