% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.build_pfam_landscape_query}
\alias{.build_pfam_landscape_query}
\title{Build Pfam landscape query}
\usage{
.build_pfam_landscape_query(blast_param_id)
}
\description{
Build Pfam landscape query
}
\keyword{internal}
