% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.compile_annotations_stage}
\alias{.compile_annotations_stage}
\title{Handle annotations stage}
\usage{
.compile_annotations_stage(
  con,
  include,
  candidate_loci,
  blast_param_id,
  verbose
)
}
\description{
Handle annotations stage
}
\keyword{internal}
