% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.compile_descriptive_stage}
\alias{.compile_descriptive_stage}
\title{Handle descriptive stage (placeholder)}
\usage{
.compile_descriptive_stage(con, data, verbose)
}
\description{
Handle descriptive stage (placeholder)
}
\keyword{internal}
