% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.compile_enrichment_stage}
\alias{.compile_enrichment_stage}
\title{Handle enrichment stage}
\usage{
.compile_enrichment_stage(
  con,
  data,
  analysis_ids,
  significance_threshold,
  verbose
)
}
\description{
Handle enrichment stage
}
\keyword{internal}
