% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.convert_clusterprofiler_to_funseqr}
\alias{.convert_clusterprofiler_to_funseqr}
\title{Convert clusterProfiler results to funseqR format}
\usage{
.convert_clusterprofiler_to_funseqr(
  clusterprofiler_result,
  ontology,
  significance_threshold,
  verbose
)
}
\description{
Convert clusterProfiler results to funseqR format
}
\keyword{internal}
