% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.convert_go_data_to_clusterprofiler}
\alias{.convert_go_data_to_clusterprofiler}
\title{Convert go_data to clusterProfiler format}
\usage{
.convert_go_data_to_clusterprofiler(go_data, ontology, verbose)
}
\description{
Convert go_data to clusterProfiler format
}
\keyword{internal}
