% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.convert_kegg_data_to_clusterprofiler}
\alias{.convert_kegg_data_to_clusterprofiler}
\title{Convert kegg_data to clusterProfiler format}
\usage{
.convert_kegg_data_to_clusterprofiler(kegg_data, verbose)
}
\description{
Convert kegg_data to clusterProfiler format
}
\keyword{internal}
