% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.create_empty_annotation_dataframe}
\alias{.create_empty_annotation_dataframe}
\title{Create empty annotation dataframe with all expected columns}
\usage{
.create_empty_annotation_dataframe()
}
\description{
Create empty annotation dataframe with all expected columns
}
\keyword{internal}
