% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_workflow.R
\name{.create_enriched_loci_summary}
\alias{.create_enriched_loci_summary}
\title{Create summary statistics for enriched loci}
\usage{
.create_enriched_loci_summary(enriched_loci, enriched_terms_list, all_loci)
}
\description{
Create summary statistics for enriched loci
}
\keyword{internal}
