% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterprofiler_enrichment.R
\name{.create_enrichment_summary}
\alias{.create_enrichment_summary}
\title{Create enrichment analysis summary}
\usage{
.create_enrichment_summary(
  results,
  candidate_genes,
  background_genes,
  candidate_file_id,
  background_file_id,
  blast_param_id
)
}
\description{
Create enrichment analysis summary
}
\keyword{internal}
