% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.create_heatmap_landscape}
\alias{.create_heatmap_landscape}
\title{Create heatmap landscape visualization}
\usage{
.create_heatmap_landscape(combined_data, color_palette)
}
\description{
Create heatmap landscape visualization
}
\keyword{internal}
