% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_workflow.R
\name{.create_stored_locus_info_table}
\alias{.create_stored_locus_info_table}
\title{Create locus information table for stored candidates ORA results}
\usage{
.create_stored_locus_info_table(con, annotation_type, verbose)
}
\description{
Create locus information table for stored candidates ORA results
}
\keyword{internal}
