% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.create_treemap_landscape}
\alias{.create_treemap_landscape}
\title{Create treemap landscape visualization}
\usage{
.create_treemap_landscape(combined_data, color_palette)
}
\description{
Create treemap landscape visualization
}
\keyword{internal}
