% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.enhance_kegg_pathway_names}
\alias{.enhance_kegg_pathway_names}
\title{Enhance KEGG pathway names using KEGGREST}
\usage{
.enhance_kegg_pathway_names(kegg_data, verbose)
}
\description{
Enhance KEGG pathway names using KEGGREST
}
\keyword{internal}
