% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_annotation_data}
\alias{.export_annotation_data}
\title{Export annotation data}
\usage{
.export_annotation_data(source_con, target_con, id_mappings, verbose)
}
\description{
Export annotation data
}
\keyword{internal}
