% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_blast_data}
\alias{.export_blast_data}
\title{Export BLAST data}
\usage{
.export_blast_data(source_con, target_con, id_mappings, verbose)
}
\description{
Export BLAST data
}
\keyword{internal}
