% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_candidate_data}
\alias{.export_candidate_data}
\title{Export candidate data with smart dependencies}
\usage{
.export_candidate_data(source_con, target_con, verbose)
}
\description{
Export candidate data with smart dependencies
}
\keyword{internal}
