% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_chromosome_metadata}
\alias{.export_chromosome_metadata}
\title{Export chromosome metadata}
\usage{
.export_chromosome_metadata(source_con, target_con, verbose)
}
\description{
Export chromosome metadata
}
\keyword{internal}
