% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_flanking_data}
\alias{.export_flanking_data}
\title{Export flanking sequences data}
\usage{
.export_flanking_data(source_con, target_con, id_mappings, verbose)
}
\description{
Export flanking sequences data
}
\keyword{internal}
