% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_functional_annotations}
\alias{.export_functional_annotations}
\title{Export functional annotations (GO, KEGG, etc.)}
\usage{
.export_functional_annotations(
  source_con,
  target_con,
  source_annotations,
  target_annotations,
  verbose
)
}
\description{
Export functional annotations (GO, KEGG, etc.)
}
\keyword{internal}
