% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_reference_data}
\alias{.export_reference_data}
\title{Export reference genome data}
\usage{
.export_reference_data(source_con, target_con, id_mappings, verbose)
}
\description{
Export reference genome data
}
\keyword{internal}
