% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{.export_vcf_data}
\alias{.export_vcf_data}
\title{Export VCF data and file registrations}
\usage{
.export_vcf_data(source_con, target_con, id_mappings, verbose)
}
\description{
Export VCF data and file registrations
}
\keyword{internal}
