% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.extract_all_go_annotations}
\alias{.extract_all_go_annotations}
\title{Extract all GO annotations from database}
\usage{
.extract_all_go_annotations(con, ontologies, verbose)
}
\description{
Extract all GO annotations from database
}
\keyword{internal}
