% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterprofiler_enrichment.R
\name{.extract_background_uniprot_ids}
\alias{.extract_background_uniprot_ids}
\title{Extract background UniProt IDs with GO annotations}
\usage{
.extract_background_uniprot_ids(
  con,
  background_file_id,
  blast_param_id,
  verbose = FALSE
)
}
\description{
Extract background UniProt IDs with GO annotations
}
\keyword{internal}
