% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterprofiler_enrichment.R
\name{.extract_candidate_uniprot_ids}
\alias{.extract_candidate_uniprot_ids}
\title{Extract candidate UniProt IDs by linking positions}
\usage{
.extract_candidate_uniprot_ids(
  con,
  candidate_file_id,
  background_file_id,
  blast_param_id,
  verbose = FALSE
)
}
\description{
Extract candidate UniProt IDs by linking positions
}
\keyword{internal}
