% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.extract_candidate_yvalues}
\alias{.extract_candidate_yvalues}
\title{Extract y-values for candidate loci from full VCF dataset}
\usage{
.extract_candidate_yvalues(
  con,
  candidate_coords,
  vcf_file_id,
  full_y_values,
  verbose
)
}
\description{
Extract y-values for candidate loci from full VCF dataset
}
\keyword{internal}
