% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.extract_enriched_go_terms}
\alias{.extract_enriched_go_terms}
\title{Extract enriched GO terms from ORA results}
\usage{
.extract_enriched_go_terms(
  con,
  analysis_ids,
  significance_threshold,
  ontologies,
  verbose
)
}
\description{
Extract enriched GO terms from ORA results
}
\keyword{internal}
