% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.extract_enrichment_data}
\alias{.extract_enrichment_data}
\title{Extract enrichment data from database}
\usage{
.extract_enrichment_data(con, analysis_ids, significance_threshold, verbose)
}
\description{
Extract enrichment data from database
}
\keyword{internal}
