% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.extract_pathway_number}
\alias{.extract_pathway_number}
\title{Extract pathway number from KEGG ID}
\usage{
.extract_pathway_number(kegg_id)
}
\description{
Extract pathway number from KEGG ID
}
\keyword{internal}
