% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterprofiler_enrichment.R
\name{.filter_go_data_by_ontology}
\alias{.filter_go_data_by_ontology}
\title{Filter GO data by ontology}
\usage{
.filter_go_data_by_ontology(go_data, ontology)
}
\description{
Filter GO data by ontology
}
\keyword{internal}
