% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_workflow.R
\name{.filter_loci_by_genes}
\alias{.filter_loci_by_genes}
\title{Filter loci by gene IDs}
\usage{
.filter_loci_by_genes(locus_info, enriched_genes, verbose)
}
\description{
Filter loci by gene IDs
}
\keyword{internal}
