% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.filter_to_requested_loci}
\alias{.filter_to_requested_loci}
\title{Filter comprehensive results to only include requested loci}
\usage{
.filter_to_requested_loci(result_data, loci_coords, verbose)
}
\description{
Filter comprehensive results to only include requested loci
}
\keyword{internal}
