% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.find_best_slim_match}
\alias{.find_best_slim_match}
\title{Find best GO Slim match for a GO term}
\usage{
.find_best_slim_match(go_term, go_category, go_slim_mapping)
}
\description{
Find best GO Slim match for a GO term
}
\keyword{internal}
