% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.find_brite_classification}
\alias{.find_brite_classification}
\title{Find BRITE classification for pathway}
\usage{
.find_brite_classification(pathway_num, brite_hierarchy)
}
\description{
Find BRITE classification for pathway
}
\keyword{internal}
