% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.format_methods_text}
\alias{.format_methods_text}
\title{Format methods information for scientific papers}
\usage{
.format_methods_text(con)
}
\description{
Format methods information for scientific papers
}
\keyword{internal}
