% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.format_revigo_output}
\alias{.format_revigo_output}
\title{Format GO data for ReviGO output}
\usage{
.format_revigo_output(go_data, include_pvalues, verbose)
}
\description{
Format GO data for ReviGO output
}
\keyword{internal}
