% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.get_annotation_landscape_data}
\alias{.get_annotation_landscape_data}
\title{Get annotation landscape data for visualization}
\usage{
.get_annotation_landscape_data(
  con,
  annotation_type,
  blast_param_id,
  top_n,
  min_genes,
  verbose
)
}
\description{
Get annotation landscape data for visualization
}
\keyword{internal}
