% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.get_kegg_brite_hierarchy}
\alias{.get_kegg_brite_hierarchy}
\title{Get KEGG BRITE hierarchy data}
\usage{
.get_kegg_brite_hierarchy()
}
\description{
Get KEGG BRITE hierarchy data
}
\keyword{internal}
