% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.get_methods_summary}
\alias{.get_methods_summary}
\title{Get methods summaries for reproducibility}
\usage{
.get_methods_summary(con)
}
\description{
Get methods summaries for reproducibility
}
\keyword{internal}
