% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.identify_candidate_loci}
\alias{.identify_candidate_loci}
\title{Identify candidate loci from various input formats}
\usage{
.identify_candidate_loci(con, candidate_input, verbose = FALSE)
}
\description{
Identify candidate loci from various input formats
}
\keyword{internal}
