% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.map_enriched_terms_to_loci}
\alias{.map_enriched_terms_to_loci}
\title{Map enriched terms to individual loci}
\usage{
.map_enriched_terms_to_loci(con, data, enrichment_data, verbose)
}
\description{
Map enriched terms to individual loci
}
\keyword{internal}
