% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.map_go_terms_to_slim}
\alias{.map_go_terms_to_slim}
\title{Map GO terms to GO Slim categories}
\usage{
.map_go_terms_to_slim(go_data, go_slim_mapping, verbose)
}
\description{
Map GO terms to GO Slim categories
}
\keyword{internal}
