% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{.map_kegg_to_brite}
\alias{.map_kegg_to_brite}
\title{Map KEGG pathways to BRITE hierarchy}
\usage{
.map_kegg_to_brite(kegg_data, brite_hierarchy, hierarchy_level, verbose)
}
\description{
Map KEGG pathways to BRITE hierarchy
}
\keyword{internal}
