% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.match_coordinates_to_loci}
\alias{.match_coordinates_to_loci}
\title{Match coordinate data frame to database loci}
\usage{
.match_coordinates_to_loci(con, coords_df, verbose = FALSE)
}
\description{
Match coordinate data frame to database loci
}
\keyword{internal}
