% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.merge_enrichment_with_data}
\alias{.merge_enrichment_with_data}
\title{Merge enrichment data with original data frame}
\usage{
.merge_enrichment_with_data(data, enriched_loci_data, verbose)
}
\description{
Merge enrichment data with original data frame
}
\keyword{internal}
