% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.parse_bed_file}
\alias{.parse_bed_file}
\title{Parse BED file into coordinate format}
\usage{
.parse_bed_file(bed_file, verbose)
}
\description{
Parse BED file into coordinate format
}
\keyword{internal}
