% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.parse_loci_input}
\alias{.parse_loci_input}
\title{Parse loci input into standardized coordinate format}
\usage{
.parse_loci_input(loci, verbose)
}
\description{
Parse loci input into standardized coordinate format
}
\keyword{internal}
