% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.perform_clusterprofiler_enrichment}
\alias{.perform_clusterprofiler_enrichment}
\title{Perform GO enrichment using clusterProfiler}
\usage{
.perform_clusterprofiler_enrichment(
  go_data,
  ontology,
  min_genes,
  max_genes,
  significance_threshold,
  verbose
)
}
\description{
Perform GO enrichment using clusterProfiler
}
\keyword{internal}
