% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.perform_clusterprofiler_kegg_enrichment}
\alias{.perform_clusterprofiler_kegg_enrichment}
\title{Perform KEGG enrichment using clusterProfiler}
\usage{
.perform_clusterprofiler_kegg_enrichment(
  kegg_data,
  min_genes,
  max_genes,
  significance_threshold,
  verbose
)
}
\description{
Perform KEGG enrichment using clusterProfiler
}
\keyword{internal}
