% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{.perform_go_enrichment_from_data}
\alias{.perform_go_enrichment_from_data}
\title{Perform GO enrichment analysis from prepared data}
\usage{
.perform_go_enrichment_from_data(
  go_data,
  ontology,
  min_genes,
  max_genes,
  significance_threshold,
  verbose
)
}
\description{
Perform GO enrichment analysis from prepared data
}
\keyword{internal}
