% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{.perform_kegg_enrichment_from_data}
\alias{.perform_kegg_enrichment_from_data}
\title{Perform KEGG enrichment analysis from prepared data}
\usage{
.perform_kegg_enrichment_from_data(
  kegg_data,
  min_pathways,
  max_pathways,
  significance_threshold,
  verbose
)
}
\description{
Perform KEGG enrichment analysis from prepared data
}
\keyword{internal}
