% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.perform_legacy_enrichment}
\alias{.perform_legacy_enrichment}
\title{Perform GO enrichment using legacy hypergeometric method}
\usage{
.perform_legacy_enrichment(
  go_data,
  ontology,
  min_genes,
  max_genes,
  significance_threshold,
  verbose
)
}
\description{
Perform GO enrichment using legacy hypergeometric method
}
\keyword{internal}
