% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterprofiler_enrichment.R
\name{.prepare_clusterprofiler_go_data}
\alias{.prepare_clusterprofiler_go_data}
\title{Prepare GO annotation data for clusterProfiler}
\usage{
.prepare_clusterprofiler_go_data(
  con,
  background_file_id,
  blast_param_id,
  verbose = FALSE
)
}
\description{
Prepare GO annotation data for clusterProfiler
}
\keyword{internal}
