% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{.prepare_go_data_from_annotations}
\alias{.prepare_go_data_from_annotations}
\title{Prepare GO data from processed annotations}
\usage{
.prepare_go_data_from_annotations(annotations, candidate_locus_ids, verbose)
}
\description{
Prepare GO data from processed annotations
}
\keyword{internal}
