% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.process_eggnog_annotations}
\alias{.process_eggnog_annotations}
\title{Process eggNOG annotations for loci}
\usage{
.process_eggnog_annotations(con, result_data, base_where, params, verbose)
}
\description{
Process eggNOG annotations for loci
}
\keyword{internal}
