% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.process_go_annotations}
\alias{.process_go_annotations}
\title{Process GO annotations for loci}
\usage{
.process_go_annotations(con, result_data, base_where, params, verbose)
}
\description{
Process GO annotations for loci
}
\keyword{internal}
