% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.process_interpro_annotations}
\alias{.process_interpro_annotations}
\title{Process InterPro annotations for loci}
\usage{
.process_interpro_annotations(con, result_data, base_where, params, verbose)
}
\description{
Process InterPro annotations for loci
}
\keyword{internal}
