% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.process_kegg_annotations}
\alias{.process_kegg_annotations}
\title{Process KEGG annotations with enhancements}
\usage{
.process_kegg_annotations(con, result_data, base_where, params, verbose)
}
\description{
Process KEGG annotations with enhancements
}
\keyword{internal}
