% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_annotations.R
\name{.process_pfam_annotations}
\alias{.process_pfam_annotations}
\title{Process Pfam annotations for loci}
\usage{
.process_pfam_annotations(con, result_data, base_where, params, verbose)
}
\description{
Process Pfam annotations for loci
}
\keyword{internal}
