% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_traceability.R
\name{.query_go_terms_to_loci}
\alias{.query_go_terms_to_loci}
\title{Query database to trace GO terms to genomic loci}
\usage{
.query_go_terms_to_loci(con, go_ids, candidate_file_id)
}
\arguments{
\item{con}{Database connection}

\item{go_ids}{Character vector of GO term IDs}

\item{candidate_file_id}{Integer. File ID of candidate dataset}
}
\value{
Data frame with genomic coordinates and functional annotations
}
\description{
Query database to trace GO terms to genomic loci
}
\keyword{internal}
