% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_traceability.R
\name{.summarize_loci_per_term}
\alias{.summarize_loci_per_term}
\title{Create summary of loci per GO term}
\usage{
.summarize_loci_per_term(traced_data)
}
\arguments{
\item{traced_data}{Data frame from trace analysis}
}
\value{
Data frame summarizing loci counts per GO term
}
\description{
Create summary of loci per GO term
}
\keyword{internal}
