% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.write_revigo_file}
\alias{.write_revigo_file}
\title{Write ReviGO-compatible file}
\usage{
.write_revigo_file(revigo_data, output_file, include_pvalues, verbose)
}
\description{
Write ReviGO-compatible file
}
\keyword{internal}
