% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{ensure_blast_db_metadata_table}
\alias{ensure_blast_db_metadata_table}
\title{Ensure the blast_database_metadata table exists}
\usage{
ensure_blast_db_metadata_table(con, verbose = FALSE)
}
\arguments{
\item{con}{A database connection object}

\item{verbose}{Logical. If TRUE, print progress information. Default is FALSE}
}
\value{
Invisible NULL
}
\description{
Ensure the blast_database_metadata table exists
}
\keyword{internal}
