% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reports.R
\name{ensure_reports_table}
\alias{ensure_reports_table}
\title{Ensure Reports Table Exists}
\usage{
ensure_reports_table(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Ensure Reports Table Exists
}
