% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{ensure_uniprot_cache_table}
\alias{ensure_uniprot_cache_table}
\title{Ensure the uniprot_cache table exists}
\usage{
ensure_uniprot_cache_table(con, verbose = FALSE)
}
\arguments{
\item{con}{A database connection object}

\item{verbose}{Logical. If TRUE, print progress information}
}
\value{
Invisible NULL
}
\description{
Ensure the uniprot_cache table exists
}
