% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_flanking_sequences.R
\name{extract_flanking_sequences}
\alias{extract_flanking_sequences}
\title{Extract flanking sequences for SNPs}
\usage{
extract_flanking_sequences(snp_positions, reference_genome, flank_size = 10000)
}
\arguments{
\item{snp_positions}{Either a character string specifying the path to a .bed file,
or a data frame with at least 3 columns: chromosome, start, and end positions.}

\item{reference_genome}{A DNAStringSet object containing the reference genome sequences.}

\item{flank_size}{The size of the flanking region to extract on each side of the SNP (default: 10000).}
}
\value{
A DNAStringSet object containing the flanking sequences for each SNP.
}
\description{
This function extracts flanking sequences for SNPs from a reference genome.
SNP positions can be provided either as a .bed file or a data frame.
If a SNP is closer to the chromosome end than the specified flank size,
the flanking sequence will be truncated to fit within the chromosome.
}
