% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reports.R
\name{generate_standard_template}
\alias{generate_standard_template}
\title{Generate Standard Template Content}
\usage{
generate_standard_template(project_info)
}
\arguments{
\item{project_info}{Project information}
}
\value{
Character vector of template content
}
\description{
Generate Standard Template Content
}
