% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reports.R
\name{generate_template_content}
\alias{generate_template_content}
\title{Generate Template Content Based on Type}
\usage{
generate_template_content(project_info, template_type)
}
\arguments{
\item{project_info}{Project information}

\item{template_type}{Template type ("minimal", "standard", "detailed")}
}
\value{
Character vector of template content
}
\description{
Generate Template Content Based on Type
}
