% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_summary.R
\name{get_blast_database_info}
\alias{get_blast_database_info}
\title{New function to get detailed BLAST database information}
\usage{
get_blast_database_info(con, db_name = NULL, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object}

\item{db_name}{Optional. Filter by specific database name}

\item{verbose}{Logical. Print detailed output. Default is TRUE}
}
\value{
A list containing detailed database information
}
\description{
New function to get detailed BLAST database information
}
