% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{get_blast_results}
\alias{get_blast_results}
\title{Get BLAST results from the database}
\usage{
get_blast_results(
  con,
  blast_param_id,
  e_value_threshold = NULL,
  max_hits_per_query = NULL,
  include_vcf_info = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{The ID of the BLAST parameters.}

\item{e_value_threshold}{Optional. E-value threshold for filtering results. Default is NULL.}

\item{max_hits_per_query}{Optional. Maximum number of hits to return per query. Default is NULL.}

\item{include_vcf_info}{Logical. If TRUE, include VCF information. Default is TRUE.}
}
\value{
A data frame containing the BLAST results.
}
\description{
This function retrieves BLAST results from the database.
}
\keyword{internal}
