% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{get_project}
\alias{get_project}
\title{Get information about a specific project}
\usage{
get_project(con, project_id)
}
\arguments{
\item{con}{A database connection object.}

\item{project_id}{The ID of the project.}
}
\value{
A data frame containing information about the project.
}
\description{
Get information about a specific project
}
