% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{get_project_id}
\alias{get_project_id}
\title{Get the ID of a project by name}
\usage{
get_project_id(con, project_name)
}
\arguments{
\item{con}{A database connection object.}

\item{project_name}{The name of the project.}
}
\value{
The ID of the project, or NULL if no project with the given name exists.
}
\description{
Get the ID of a project by name
}
