% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reports.R
\name{get_report_info}
\alias{get_report_info}
\title{Get Report Information from Database}
\usage{
get_report_info(con, project_id)
}
\arguments{
\item{con}{Database connection}

\item{project_id}{Project ID}
}
\value{
List with report information or NULL
}
\description{
Get Report Information from Database
}
