% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_schema.R
\name{get_schema_version}
\alias{get_schema_version}
\title{Check the schema version of a funseqR database}
\usage{
get_schema_version(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A numeric value representing the schema version.
}
\description{
Check the schema version of a funseqR database
}
