% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_summary.R
\name{get_table_counts}
\alias{get_table_counts}
\title{Quick table counts for all tables}
\usage{
get_table_counts(con, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object}

\item{verbose}{Logical. Print output. Default is TRUE}
}
\value{
Named vector of table counts
}
\description{
Quick table counts for all tables
}
