% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{get_vcf_by_position}
\alias{get_vcf_by_position}
\title{Get VCF data by chromosome and position}
\usage{
get_vcf_by_position(con, file_id, chromosome, start_pos, end_pos)
}
\arguments{
\item{con}{A database connection object.}

\item{file_id}{The ID of the input file containing the VCF data.}

\item{chromosome}{The chromosome to filter by.}

\item{start_pos}{The start position of the range.}

\item{end_pos}{The end position of the range.}
}
\value{
A data frame containing the VCF data.
}
\description{
This function retrieves VCF data for a specific chromosome and position range.
}
