% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{import_blast_results}
\alias{import_blast_results}
\title{Import BLAST results into the database}
\usage{
import_blast_results(
  con,
  blast_param_id,
  results_file,
  vcf_file_id,
  verbose = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{The ID of the BLAST parameters.}

\item{results_file}{Path to the BLAST results file.}

\item{vcf_file_id}{The VCF file ID for looking up flanking sequences.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
The number of BLAST results imported.
}
\description{
This function imports BLAST results from a file into the database with proper
mapping between query sequence IDs and flanking sequence IDs.
}
\keyword{internal}
