% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sequence.R
\name{import_flanking_seqs_to_db}
\alias{import_flanking_seqs_to_db}
\title{Import flanking sequences into the database}
\usage{
import_flanking_seqs_to_db(
  con,
  vcf_file_id,
  genome_id,
  flank_size = 10000,
  chromosome = NULL,
  verbose = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{vcf_file_id}{The ID of the input file containing the VCF data.}

\item{genome_id}{The ID of the reference genome to use.}

\item{flank_size}{The size of the flanking region to extract on each side of the SNP. Default is 10000.}

\item{chromosome}{Optional. Limit extraction to a specific chromosome. Default is NULL (all chromosomes).}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A list containing:
\item{vcf_count}{The number of VCF entries processed.}
\item{flanking_count}{The number of flanking sequences extracted.}
}
\description{
This function extracts flanking sequences for VCF data in the database and stores them.
It can be used as an alternative to the original extract_flanking_sequences function.
}
