% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{import_reference}
\alias{import_reference}
\title{Sequence data storage and retrieval functions for funseqR}
\usage{
import_reference(
  con,
  genome_file,
  genome_name = NULL,
  genome_build = NULL,
  store_sequences = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{genome_file}{A character string specifying the path to the FASTA file.}

\item{genome_name}{A character string specifying a name for the genome. Default is NULL (derived from file name).}

\item{genome_build}{A character string specifying the build version of the genome. Default is NULL.}

\item{store_sequences}{Logical. If TRUE, store the actual sequence data in the database. Default is TRUE.
If FALSE, only sequence metadata will be stored, and the original file will need to be accessible.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A list containing:
\item{file_id}{The ID of the registered input file.}
\item{genome_id}{The ID of the imported genome.}
\item{sequence_count}{The number of sequences imported.}
}
\description{
These functions manage reference genomes and flanking sequences within the funseqR database.
}
\details{
Import a reference genome into the database

This function reads a reference genome file (FASTA) and stores its contents in the database.
}
