% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{import_vcf}
\alias{import_vcf}
\title{VCF data storage and retrieval functions for funseqR}
\usage{
import_vcf(con, vcf_file, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{vcf_file}{A character string specifying the path to the VCF file.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A list containing:
\item{file_id}{The ID of the registered input file.}
\item{vcf_count}{The number of VCF entries imported.}
}
\description{
These functions manage VCF data within the funseqR database.
}
\details{
Import a VCF file into the database

This function reads a VCF file and stores its contents in the database.
It can be used as an alternative to the original read_vcf function.
}
