% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{list_blast_params}
\alias{list_blast_params}
\title{List all BLAST parameters in the database}
\usage{
list_blast_params(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame containing all BLAST parameters in the database.
}
\description{
List all BLAST parameters in the database
}
