% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{list_projects}
\alias{list_projects}
\title{Get a list of projects in the database}
\usage{
list_projects(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame containing information about all projects in the database.
}
\description{
Get a list of projects in the database
}
