% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterprofiler_enrichment.R
\name{print.funseqR_enrichment}
\alias{print.funseqR_enrichment}
\title{Print method for funseqR enrichment results}
\usage{
\method{print}{funseqR_enrichment}(x, ...)
}
\description{
Print method for funseqR enrichment results
}
