% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{query_uniprot_api}
\alias{query_uniprot_api}
\title{Query the UniProt API for a protein accession}
\usage{
query_uniprot_api(
  accession,
  fields = "accession,id,gene_names,go,xref_kegg,organism_name,protein_name",
  debug = FALSE
)
}
\arguments{
\item{accession}{The UniProt accession number}

\item{fields}{The fields to retrieve from UniProt (for search endpoint)}

\item{debug}{If TRUE, save the API response to a file for debugging}
}
\value{
A list containing API response with consistent structure
}
\description{
Query the UniProt API for a protein accession
}
