% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{read_uniprot_json}
\alias{read_uniprot_json}
\title{Read UniProt JSON data from a file and extract annotations}
\usage{
read_uniprot_json(accession, json_file = NULL, debug = FALSE)
}
\arguments{
\item{accession}{The UniProt accession number}

\item{json_file}{The path to the JSON file (if NULL, will look in uniprot_debug directory)}

\item{debug}{If TRUE, print debugging information}
}
\value{
A list containing extracted annotation information
}
\description{
This function reads a UniProt JSON file and extracts annotation information
}
