% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read a VCF file}
\usage{
read_vcf(vcf_file, verbose = FALSE)
}
\arguments{
\item{vcf_file}{A character string specifying the path to the VCF file.}

\item{verbose}{Logical. If TRUE, progress information is printed during file reading.}
}
\value{
A vcfR object containing the VCF data.
}
\description{
This function reads a VCF (Variant Call Format) file into R using the vcfR package.
}
\examples{
\dontrun{
vcf_data <- read_vcf("path/to/your/file.vcf")
}

}
