% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{store_blast_db_metadata}
\alias{store_blast_db_metadata}
\title{Store BLAST database metadata in the database}
\usage{
store_blast_db_metadata(con, blast_param_id, metadata, verbose = FALSE)
}
\arguments{
\item{con}{A database connection object}

\item{blast_param_id}{The ID of the BLAST parameters to associate with this metadata}

\item{metadata}{A list containing database metadata as returned by extract_blast_db_metadata}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE}
}
\value{
The ID of the stored metadata record, or NULL if storage fails
}
\description{
Store BLAST database metadata in the database
}
\keyword{internal}
