% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{store_go_enrichment_results}
\alias{store_go_enrichment_results}
\title{Store GO enrichment analysis results in the database}
\usage{
store_go_enrichment_results(
  con,
  foreground_file_id,
  background_file_id,
  enrichment_results,
  ontology,
  parameters = NULL,
  verbose = TRUE
)
}
\arguments{
\item{con}{Database connection object}

\item{foreground_file_id}{Integer. File ID of foreground dataset}

\item{background_file_id}{Integer. File ID of background dataset}

\item{enrichment_results}{Data frame. Results from perform_go_enrichment()}

\item{ontology}{Character. GO ontology tested}

\item{parameters}{List. Analysis parameters for reproducibility}

\item{verbose}{Logical. Print progress information. Default is TRUE}
}
\value{
Integer. The enrichment_id of the stored analysis
}
\description{
Store GO enrichment analysis results in the database
}
