% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{store_main_chromosomes}
\alias{store_main_chromosomes}
\title{Store main chromosomes in database metadata}
\usage{
store_main_chromosomes(con, chromosomes)
}
\arguments{
\item{con}{Database connection}

\item{chromosomes}{Character vector of chromosome names}
}
\description{
Store main chromosomes in database metadata
}
