% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reports.R
\name{store_report_info}
\alias{store_report_info}
\title{Store Report Information in Database}
\usage{
store_report_info(con, project_id, report_path, format, template)
}
\arguments{
\item{con}{Database connection}

\item{project_id}{Project ID}

\item{report_path}{Report path}

\item{format}{Report format}

\item{template}{Template type}
}
\description{
Store Report Information in Database
}
