% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{test_uniprot_connection}
\alias{test_uniprot_connection}
\title{Test connection to the UniProt API using working endpoints}
\usage{
test_uniprot_connection(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. If TRUE, print progress information}
}
\value{
Logical. TRUE if the connection is successful
}
\description{
Test connection to the UniProt API using working endpoints
}
