% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_schema.R
\name{upgrade_funseq_schema}
\alias{upgrade_funseq_schema}
\title{Upgrade the funseqR database schema}
\usage{
upgrade_funseq_schema(con, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
Invisible NULL.
}
\description{
This function upgrades the schema of an existing funseqR database to the current version.
}
