% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{vcf_to_bed}
\alias{vcf_to_bed}
\title{Convert VCF data in the database to BED format}
\usage{
vcf_to_bed(con, file_id, output_file = NULL, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{file_id}{The ID of the input file containing the VCF data.}

\item{output_file}{Optional. The file path to the output BED file. If not supplied, the function will only return the data frame.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A data frame with columns chromosome, start_position, and end_position.
}
\description{
This function generates BED format data from VCF data stored in the database.
It can be used as an alternative to the original vcf2bed function.
}
