% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{verify_annotations}
\alias{verify_annotations}
\title{Verify annotation storage for processed entries}
\usage{
verify_annotations(con, annotation_ids, verbose = FALSE)
}
\arguments{
\item{con}{Database connection object}

\item{annotation_ids}{Vector of annotation IDs to verify}

\item{verbose}{Print debug information}
}
\value{
A list with verification results
}
\description{
Verify annotation storage for processed entries
}
