% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{annotate_blast_results}
\alias{annotate_blast_results}
\title{Process BLAST results and retrieve UniProt annotations}
\usage{
annotate_blast_results(
  con,
  blast_param_id,
  max_hits = 5,
  e_value_threshold = 1e-10,
  batch_size = 500,
  delay = 1,
  offline_mode = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{The ID of the BLAST parameters.}

\item{max_hits}{Maximum number of hits to process per query. Default is 5.}

\item{e_value_threshold}{E-value threshold for filtering BLAST hits. Default is 1e-10.}

\item{batch_size}{Number of annotations to process in each transaction. Default is 500.}

\item{delay}{Delay between API calls in seconds. Default is 1.}

\item{offline_mode}{If TRUE, skips UniProt API and uses basic annotations. Default is FALSE.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A list containing annotation statistics.
}
\description{
This function processes BLAST results from the database and retrieves
UniProt annotations for the hit accessions.
}
