% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{begin_transaction}
\alias{begin_transaction}
\title{Begin a transaction in the database}
\usage{
begin_transaction(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
Invisible NULL.
}
\description{
Begin a transaction in the database
}
