% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\name{commit_transaction}
\alias{commit_transaction}
\title{Commit a transaction in the database}
\usage{
commit_transaction(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
Invisible NULL.
}
\description{
Commit a transaction in the database
}
\examples{
\dontrun{
con <- connect_funseq_db("analysis.db")
begin_transaction(con)
# ... perform database operations ...
commit_transaction(con)
}

}
