% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{commit_transaction}
\alias{commit_transaction}
\title{Commit a transaction in the database}
\usage{
commit_transaction(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
Invisible NULL.
}
\description{
Commit a transaction in the database
}
