% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vcf.R
\name{consolidate_scaffolds}
\alias{consolidate_scaffolds}
\title{Consolidate scaffolds into a pseudo-chromosome for visualization}
\usage{
consolidate_scaffolds(
  vcf_data,
  keep_chromosomes,
  pseudo_name = "US",
  verbose = FALSE
)
}
\arguments{
\item{vcf_data}{Data frame. VCF data with at least a 'chromosome' column}

\item{keep_chromosomes}{Character vector. Chromosome names to keep separate}

\item{pseudo_name}{Character. Name for the consolidated pseudo-chromosome. Default is "US"}

\item{verbose}{Logical. If TRUE, print consolidation information. Default is TRUE}
}
\value{
Data frame. VCF data with consolidated chromosome names
}
\description{
This function takes VCF data and consolidates small scaffolds into a single
pseudo-chromosome called "US" (unmapped scaffolds) while retaining specified
main chromosomes. This is useful for visualization and analysis when you have
many small scaffolds that clutter plots.
}
\details{
All chromosomes not specified in \code{keep_chromosomes} will be renamed to the
\code{pseudo_name} value. The original data is not modified - a new data frame
is returned with updated chromosome names.
}
\examples{
\dontrun{
# Get VCF data from database
vcf_data <- get_vcf_data(con, file_id = 1)

# Consolidate all except main linkage groups
main_chroms <- c("LG1", "LG2", "LG3", "LG4", "LG5")
consolidated <- consolidate_scaffolds(vcf_data, main_chroms)

# Check the result
table(consolidated$chromosome)
}

}
