% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{count_annotations}
\alias{count_annotations}
\title{Count annotations in the database}
\usage{
count_annotations(con, blast_param_id = NULL)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{Optional. The ID of the BLAST parameters. Default is NULL.}
}
\value{
A list containing counts of annotations, GO terms, and KEGG references.
}
\description{
Count annotations in the database
}
\keyword{internal}
