% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_blast.R
\name{count_blast_results}
\alias{count_blast_results}
\title{Count BLAST results in the database}
\usage{
count_blast_results(con, blast_param_id = NULL)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{Optional. The ID of the BLAST parameters. Default is NULL.}
}
\value{
The number of BLAST results.
}
\description{
Count BLAST results in the database
}
