% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vcf.R
\name{count_vcf_entries}
\alias{count_vcf_entries}
\title{Get a count of VCF entries in the database}
\usage{
count_vcf_entries(con, file_id = NULL, project_id = NULL)
}
\arguments{
\item{con}{A database connection object.}

\item{file_id}{The ID of the input file. If NULL, count all entries. Default is NULL.}

\item{project_id}{The ID of the project. If NULL, count all entries. Default is NULL.}
}
\value{
The number of VCF entries.
}
\description{
Get a count of VCF entries in the database
}
