% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{create_empty_annotation}
\alias{create_empty_annotation}
\title{Create an empty annotation structure}
\usage{
create_empty_annotation(accession)
}
\arguments{
\item{accession}{The accession number}
}
\value{
A list with empty annotation structures
}
\description{
Create an empty annotation structure
}
