% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{create_funseq_db}
\alias{create_funseq_db}
\title{Database connection utilities for funseqR}
\usage{
create_funseq_db(db_path, force = FALSE, verbose = TRUE)
}
\arguments{
\item{db_path}{Character string specifying the path to the database file.
If the file already exists, it will not be overwritten unless force=TRUE.}

\item{force}{Logical. If TRUE, the existing database will be overwritten. Default is FALSE.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A connection object to the newly created database.
}
\description{
These functions manage connections to the SQLite database that stores funseqR data.
}
\details{
Create a new funseqR database

This function creates a new SQLite database for storing funseqR data.
It sets up the required schema and initializes the database.
}
