% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sequence.R
\name{delete_flanking_sequences}
\alias{delete_flanking_sequences}
\title{Delete flanking sequences from the database}
\usage{
delete_flanking_sequences(con, vcf_file_id, confirm = TRUE, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{vcf_file_id}{The ID of the input file containing the VCF data.}

\item{confirm}{Logical. If TRUE, user confirmation will be required. Default is TRUE.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
Logical. TRUE if the deletion was successful.
}
\description{
Delete flanking sequences from the database
}
