% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vcf.R
\name{delete_vcf_data}
\alias{delete_vcf_data}
\title{Delete VCF data from the database}
\usage{
delete_vcf_data(con, file_id, confirm = TRUE, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{file_id}{The ID of the input file containing the VCF data to delete.}

\item{confirm}{Logical. If TRUE, user confirmation will be required. Default is TRUE.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
Logical. TRUE if the deletion was successful.
}
\description{
Delete VCF data from the database
}
