% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_workflow.R
\name{.create_locus_info_table}
\alias{.create_locus_info_table}
\title{Create locus information table for ORA results}
\usage{
.create_locus_info_table(
  con,
  candidate_file_id,
  background_file_id,
  blast_param_id,
  annotation_type,
  verbose
)
}
\description{
Create locus information table for ORA results
}
\keyword{internal}
