% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.extract_candidate_go_annotations}
\alias{.extract_candidate_go_annotations}
\title{Extract GO annotations for candidate loci}
\usage{
.extract_candidate_go_annotations(
  con,
  candidate_loci,
  candidate_locus_pvalues,
  include_pvalues,
  ontologies,
  verbose
)
}
\description{
Extract GO annotations for candidate loci
}
\keyword{internal}
