% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{.extract_stored_candidate_kegg_annotations}
\alias{.extract_stored_candidate_kegg_annotations}
\title{Extract KEGG annotations for stored candidates}
\usage{
.extract_stored_candidate_kegg_annotations(
  con,
  background_file_id,
  blast_param_id = NULL,
  verbose = TRUE
)
}
\description{
Extract KEGG annotations for stored candidates
}
\keyword{internal}
