% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{extract_uniprot_info}
\alias{extract_uniprot_info}
\title{Extract UniProt information from API response}
\usage{
extract_uniprot_info(uniprot_data, debug = FALSE)
}
\arguments{
\item{uniprot_data}{The response from query_uniprot_api}

\item{debug}{If TRUE, print debugging information}
}
\value{
A list containing extracted information
}
\description{
Extract UniProt information from API response
}
