% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reports.R
\name{generate_report_template}
\alias{generate_report_template}
\title{Generate Report Template}
\usage{
generate_report_template(project_info, format, template)
}
\arguments{
\item{project_info}{Project information}

\item{format}{Report format ("Rmd" or "Qmd")}

\item{template}{Template type}
}
\value{
Character vector of template content
}
\description{
Generate Report Template
}
