% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_annotations}
\alias{get_annotations}
\title{Get annotations from the database}
\usage{
get_annotations(
  con,
  blast_param_id,
  include_go = TRUE,
  include_kegg = TRUE,
  include_vcf_info = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{The ID of the BLAST parameters.}

\item{include_go}{Logical. If TRUE, include GO terms. Default is TRUE.}

\item{include_kegg}{Logical. If TRUE, include KEGG references. Default is TRUE.}

\item{include_vcf_info}{Logical. If TRUE, include VCF information. Default is TRUE.}
}
\value{
A list containing:
\item{annotations}{A data frame of annotations.}
\item{go_terms}{A data frame of GO terms (if include_go is TRUE).}
\item{kegg_refs}{A data frame of KEGG references (if include_kegg is TRUE).}
}
\description{
This function retrieves annotations from the database.
}
