% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_blast_metadata.R
\name{get_blast_db_metadata}
\alias{get_blast_db_metadata}
\title{Get BLAST database metadata from the database}
\usage{
get_blast_db_metadata(con, blast_param_id = NULL, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object}

\item{blast_param_id}{Optional. The ID of the BLAST parameters. If NULL, returns all metadata}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE}
}
\value{
A data frame containing database metadata
}
\description{
Get BLAST database metadata from the database
}
