% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_blast.R
\name{get_blast_params}
\alias{get_blast_params}
\title{Get BLAST parameters from the database}
\usage{
get_blast_params(con, blast_param_id)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{The ID of the BLAST parameters to retrieve.}
}
\value{
A data frame containing the BLAST parameters.
}
\description{
Get BLAST parameters from the database
}
