% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{get_cached_uniprot_data}
\alias{get_cached_uniprot_data}
\title{Get cached UniProt data}
\usage{
get_cached_uniprot_data(con, accession)
}
\arguments{
\item{con}{A database connection object}

\item{accession}{The UniProt accession number}
}
\value{
The cached UniProt data or NULL if not found
}
\description{
Get cached UniProt data
}
