% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{get_db_path}
\alias{get_db_path}
\title{Get the path of a database from its connection}
\usage{
get_db_path(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
Character string. The path to the database file.
}
\description{
Get the path of a database from its connection
}
