% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sequence.R
\name{get_flanking_sequences}
\alias{get_flanking_sequences}
\title{Get flanking sequences from the database}
\usage{
get_flanking_sequences(
  con,
  vcf_file_id,
  as_dna_string_set = TRUE,
  chromosome = NULL
)
}
\arguments{
\item{con}{A database connection object.}

\item{vcf_file_id}{The ID of the input file containing the VCF data.}

\item{as_dna_string_set}{Logical. If TRUE, return a DNAStringSet object. Default is TRUE.
If FALSE, return a data frame with sequence data.}

\item{chromosome}{Optional. Limit retrieval to a specific chromosome. Default is NULL (all chromosomes).}
}
\value{
A DNAStringSet object or a data frame, depending on as_dna_string_set.
}
\description{
Get flanking sequences from the database
}
