% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{get_input_file}
\alias{get_input_file}
\title{Get information about a specific file}
\usage{
get_input_file(con, file_id)
}
\arguments{
\item{con}{A database connection object.}

\item{file_id}{The ID of the file.}
}
\value{
A data frame containing information about the file.
}
\description{
Get information about a specific file
}
