% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vcf.R
\name{get_main_chromosomes}
\alias{get_main_chromosomes}
\title{Retrieve main chromosomes from database metadata}
\usage{
get_main_chromosomes(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
Character vector of main chromosome names, or NULL if not defined
}
\description{
Retrieve main chromosomes from database metadata
}
