% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sequence.R
\name{get_reference_genome}
\alias{get_reference_genome}
\title{Get a reference genome from the database}
\usage{
get_reference_genome(
  con,
  genome_id,
  as_dna_string_set = TRUE,
  include_sequences = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{genome_id}{The ID of the genome to retrieve.}

\item{as_dna_string_set}{Logical. If TRUE, return a DNAStringSet object. Default is TRUE.
If FALSE, return a data frame with sequence data.}

\item{include_sequences}{Logical. If TRUE, include the actual sequence data. Default is TRUE.
If FALSE, only include metadata.}
}
\value{
A DNAStringSet object or a data frame, depending on as_dna_string_set.
}
\description{
This function retrieves a reference genome from the database.
}
