% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vcf.R
\name{get_vcf_data}
\alias{get_vcf_data}
\title{Get VCF data from the database}
\usage{
get_vcf_data(con, file_id, limit = NULL, offset = 0)
}
\arguments{
\item{con}{A database connection object.}

\item{file_id}{The ID of the input file containing the VCF data.}

\item{limit}{The maximum number of entries to return. Default is NULL (no limit).}

\item{offset}{The number of entries to skip. Default is 0.}
}
\value{
A data frame containing the VCF data.
}
\description{
This function retrieves VCF data from the database.
}
