% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_blast.R
\name{import_blast_results}
\alias{import_blast_results}
\title{Import BLAST results into the database}
\usage{
import_blast_results(
  con,
  blast_param_id,
  results_file,
  flanking_ids,
  verbose = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_param_id}{The ID of the BLAST parameters associated with these results.}

\item{results_file}{A character string specifying the path to the BLAST results file.}

\item{flanking_ids}{A vector of flanking sequence IDs that correspond to the query sequences.
Must be in the same order as the queries in the BLAST results file.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
The number of BLAST results imported.
}
\description{
This function imports BLAST results from a tabular output file into the database.
}
