% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{is_db_connected}
\alias{is_db_connected}
\title{Check if a database connection is open}
\usage{
is_db_connected(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
Logical. TRUE if the connection is open, FALSE otherwise.
}
\description{
Check if a database connection is open
}
