% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_blast.R
\name{list_blast_params}
\alias{list_blast_params}
\title{List BLAST parameters for a project}
\usage{
list_blast_params(con, project_id)
}
\arguments{
\item{con}{A database connection object.}

\item{project_id}{The ID of the project.}
}
\value{
A data frame containing all BLAST parameters for the project.
}
\description{
List BLAST parameters for a project
}
