% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{list_input_files}
\alias{list_input_files}
\title{List all input files in the database}
\usage{
list_input_files(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame containing information about all input files in the database.
}
\description{
List all input files in the database
}
