% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_projects.R
\name{list_input_files}
\alias{list_input_files}
\title{List all input files for a project}
\usage{
list_input_files(con, project_id)
}
\arguments{
\item{con}{A database connection object.}

\item{project_id}{The ID of the project.}
}
\value{
A data frame containing information about all input files for the project.
}
\description{
List all input files for a project
}
