% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{perform_go_enrichment}
\alias{perform_go_enrichment}
\title{Perform GO enrichment analysis using hypergeometric test}
\usage{
perform_go_enrichment(
  go_data,
  ontology = "BP",
  min_genes = 5,
  max_genes = 500,
  verbose = TRUE
)
}
\arguments{
\item{go_data}{List. Output from extract_go_terms_for_enrichment()}

\item{ontology}{Character. GO ontology: "BP" (Biological Process), "MF" (Molecular Function), or "CC" (Cellular Component)}

\item{min_genes}{Integer. Minimum genes required for a GO term to be tested. Default is 5}

\item{max_genes}{Integer. Maximum genes for a GO term (to exclude very broad terms). Default is 500}

\item{verbose}{Logical. Print progress information. Default is TRUE}
}
\value{
Data frame with enrichment results, sorted by adjusted p-value
}
\description{
Perform GO enrichment analysis using hypergeometric test
}
\details{
Performs hypergeometric enrichment testing for GO terms. Tests whether each GO term
is overrepresented in the foreground set compared to the background set.
Applies FDR correction for multiple testing.
}
\examples{
\dontrun{
go_data <- extract_go_terms_for_enrichment(con, fg_id, bg_id)
bp_results <- perform_go_enrichment(go_data, "BP")
head(bp_results)
}

}
