% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{register_input_file}
\alias{register_input_file}
\title{File management functions for funseqR}
\usage{
register_input_file(
  con,
  file_path,
  file_type,
  calculate_hash = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{file_path}{Character string specifying the path to the file.}

\item{file_type}{Character string specifying the type of the file (e.g., "vcf", "fasta").}

\item{calculate_hash}{Logical. If TRUE, calculate and store a hash of the file. Default is TRUE.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
The ID of the newly registered file.
}
\description{
These functions manage input files within the funseqR database.
}
\details{
Register an input file in the database

This function adds information about an input file to the database.
}
