% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{require_uniprot_connection}
\alias{require_uniprot_connection}
\title{connection check for annotate_blast_results function}
\usage{
require_uniprot_connection(verbose = TRUE)
}
\description{
connection check for annotate_blast_results function
}
