% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{retrieve_vcf_data}
\alias{retrieve_vcf_data}
\title{Retrieve VCF data from database with optional export}
\usage{
retrieve_vcf_data(con, file_id, export_path = NULL, verbose = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{file_id}{The ID of the input file containing the VCF data.}

\item{export_path}{Optional. File path to export the VCF data. If NULL, no export is performed.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}
}
\value{
A data frame containing the VCF data from the vcf_data table.
}
\description{
This function retrieves VCF data from the database and optionally exports it to a VCF file.
Uses the stored vcfR object for fast, faithful export.
}
