% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{rollback_transaction}
\alias{rollback_transaction}
\title{Rollback a transaction in the database}
\usage{
rollback_transaction(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
Invisible NULL.
}
\description{
Rollback a transaction in the database
}
