% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{store_annotation}
\alias{store_annotation}
\title{Store annotation in the database}
\usage{
store_annotation(
  con,
  blast_result_id,
  uniprot_info,
  verbose = TRUE,
  update_existing = TRUE
)
}
\arguments{
\item{con}{A database connection object.}

\item{blast_result_id}{The ID of the BLAST result to annotate.}

\item{uniprot_info}{A list containing UniProt information as returned by extract_uniprot_info.}

\item{verbose}{Logical. If TRUE, print progress information. Default is TRUE.}

\item{update_existing}{Logical. If TRUE, add GO terms and KEGG references to existing annotations. Default is TRUE.}
}
\value{
The ID of the annotation.
}
\description{
This function stores UniProt annotation in the database or adds GO terms and KEGG references
to existing annotations if they're missing.
}
