% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{store_uniprot_data}
\alias{store_uniprot_data}
\title{Enhanced function to store UniProt API responses in database cache}
\usage{
store_uniprot_data(con, accession, response, debug = FALSE)
}
\arguments{
\item{con}{Database connection object}

\item{accession}{UniProt accession number}

\item{response}{API response (can be various formats)}

\item{debug}{Print debug information}
}
\value{
TRUE if storage successful, FALSE otherwise
}
\description{
Enhanced function to store UniProt API responses in database cache
}
