% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotation.R
\name{test_uniprot_connection}
\alias{test_uniprot_connection}
\title{Test the connection to the UniProt API}
\usage{
test_uniprot_connection()
}
\value{
Logical. TRUE if the connection is successful.
}
\description{
This function tests if the UniProt API is accessible.
}
